+++ CHANGELOG.md
@@ -17,6 +17,7 @@ require depending on default function and bootloader code
 - Support for Elixir `Enum.split_with/2`, `Enum.join/2`, `Enum.map_join/3`
 - Support for Elixir exceptions: `Exception` module and the other error related modules such as
 `ArgumentError`, `UndefinedFunctionError`, etc...
+- Support for Elixir `Enumerable` protocol
 
 ## [0.6.3] - 20-07-2024